requirements.txt
Flask>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Werkzeug>=3.1.0
WTForms>=3.1.1
scikit-learn>=1.6.0
numpy>=2.0.0
pandas>=2.2.0
joblib>=1.3.0
python-dotenv>=1.0.0
email-validator>=2.1.0
